# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ByteDock coding assessment platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /problems:
    get:
      operationId: listProblems
      tags: [problems]
      summary: List coding problems
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: difficulty
          in: query
          schema:
            type: string
            enum: [easy, medium, hard]
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, published, archived]
      responses:
        "200":
          description: Problems
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Problem"
    post:
      operationId: createProblem
      tags: [problems]
      summary: Create a coding problem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProblemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
  /problems/{id}:
    get:
      operationId: getProblem
      tags: [problems]
      summary: Get a problem
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Problem
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
        "404":
          $ref: "#/components/responses/Error"
    patch:
      operationId: updateProblem
      tags: [problems]
      summary: Update a coding problem
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProblemUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
    delete:
      operationId: deleteProblem
      tags: [problems]
      summary: Delete a coding problem
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: Deleted
    post:
      operationId: publishProblem
      tags: [problems]
      summary: Publish a coding problem
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
  /problems/{id}/test-cases:
    post:
      operationId: addTestCase
      tags: [problems]
      summary: Add a test case
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestCaseInput"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestCase"
  /execute:
    post:
      operationId: executeCode
      tags: [execution]
      summary: Execute Python code against a problem's test cases
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExecuteInput"
      responses:
        "200":
          description: Execution result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionResult"
  /submissions:
    get:
      operationId: listSubmissions
      tags: [submissions]
      summary: List submissions for the current student or staff view
      parameters:
        - name: problemId
          in: query
          schema: { type: string }
        - name: studentId
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Submissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Submission"
    post:
      operationId: createSubmission
      tags: [submissions]
      summary: Submit code for judging
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmissionInput"
      responses:
        "201":
          description: Judged submission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
  /submissions/{id}:
    get:
      operationId: getSubmission
      tags: [submissions]
      summary: Get a submission
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Submission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
...
  /students/{id}/analytics:
    get:
      operationId: getStudentAnalytics
      tags: [students]
      summary: Get student performance analytics
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Student analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentAnalytics"
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string }
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    TestCase:
      type: object
      properties:
        id: { type: string }
        input: { type: string }
        expectedOutput: { type: string }
        isSample: { type: boolean }
      required: [id, input, expectedOutput, isSample]
    TestCaseInput:
      type: object
      properties:
        input: { type: string }
        expectedOutput: { type: string }
        isSample: { type: boolean }
      required: [input, expectedOutput]
    Problem:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        slug: { type: string }
        difficulty:
          type: string
          enum: [easy, medium, hard]
        category: { type: string }
        description: { type: string }
        starterCode: { type: string }
        status:
          type: string
          enum: [draft, published, archived]
        acceptanceRate: { type: number }
        solveCount: { type: integer }
        testCases:
          type: array
          items:
            $ref: "#/components/schemas/TestCase"
        createdAt: { type: string, format: date-time }
      required: [id, title, slug, difficulty, category, description, starterCode, status, acceptanceRate, solveCount, testCases, createdAt]
    ProblemInput:
      type: object
      properties:
        title: { type: string, minLength: 3 }
        category: { type: string }
        difficulty:
          type: string
          enum: [easy, medium, hard]
        description: { type: string }
        starterCode: { type: string }
        status:
          type: string
          enum: [draft, published]
      required: [title, category, difficulty, description, starterCode]
    ProblemUpdate:
      type: object
      properties:
        title: { type: string, minLength: 3 }
        category: { type: string }
        difficulty:
          type: string
          enum: [easy, medium, hard]
        description: { type: string }
        starterCode: { type: string }
        status:
          type: string
          enum: [draft, published, archived]
    ExecuteInput:
      type: object
      properties:
        problemId: { type: string }
        code: { type: string }
      required: [problemId, code]
    ExecutionResult:
      type: object
      properties:
        status:
          type: string
          enum: [passed, failed, error, timeout]
        passed: { type: integer }
        total: { type: integer }
        runtimeMs: { type: integer }
        output: { type: string }
        testResults:
          type: array
          items:
            $ref: "#/components/schemas/TestResult"
      required: [status, passed, total, runtimeMs, output, testResults]
    TestResult:
      type: object
      properties:
        testCaseId: { type: string }
        passed: { type: boolean }
        actual: { type: string }
        expected: { type: string }
      required: [testCaseId, passed, actual, expected]
    SubmissionInput:
      type: object
      properties:
        problemId: { type: string }
        code: { type: string }
      required: [problemId, code]
    Submission:
      type: object
      properties:
        id: { type: string }
        problemId: { type: string }
        problemTitle: { type: string }
        studentId: { type: string }
        studentName: { type: string }
        code: { type: string }
        status:
          type: string
          enum: [accepted, rejected, error]
        passedTests: { type: integer }
        totalTests: { type: integer }
        runtimeMs: { type: integer }
        createdAt: { type: string, format: date-time }
      required: [id, problemId, problemTitle, studentId, studentName, code, status, passedTests, totalTests, runtimeMs, createdAt]
    StudentSummary:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string }
        solvedCount: { type: integer }
        submissionCount: { type: integer }
        lastActiveAt: { type: ["string", "null"], format: date-time }
      required: [id, name, email, solvedCount, submissionCount, lastActiveAt]
    StudentAnalytics:
      type: object
      properties:
        student:
          $ref: "#/components/schemas/StudentSummary"
        accepted: { type: integer }
        rejected: { type: integer }
        byDifficulty:
          type: array
          items:
            $ref: "#/components/schemas/Metric"
        recentSubmissions:
          type: array
          items:
            $ref: "#/components/schemas/Submission"
      required: [student, accepted, rejected, byDifficulty, recentSubmissions]
...
    Metric:
      type: object
      properties:
        label: { type: string }
        value: { type: integer }
      required: [label, value]
...