# lib/api-zod/src/generated/types/metric.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * ByteDock coding assessment platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Metric {
  label: string;
  value: number;
}