# lib/api-zod/src/generated/types/problem.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * ByteDock coding assessment platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ProblemDifficulty } from './problemDifficulty';
import type { ProblemStatus } from './problemStatus';
import type { TestCase } from './testCase';

export interface Problem {
  id: string;
  title: string;
  slug: string;
  difficulty: ProblemDifficulty;
  category: string;
  description: string;
  starterCode: string;
  status: ProblemStatus;
  acceptanceRate: number;
  solveCount: number;
  testCases: TestCase[];
  createdAt: Date;
}