# lib/api-zod/src/generated/types/problemUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * ByteDock coding assessment platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ProblemUpdateDifficulty } from './problemUpdateDifficulty';
import type { ProblemUpdateStatus } from './problemUpdateStatus';

export interface ProblemUpdate {
  /** @minLength 3 */
  title?: string;
  category?: string;
  difficulty?: ProblemUpdateDifficulty;
  description?: string;
  starterCode?: string;
  status?: ProblemUpdateStatus;
}